Add files via upload
Here is the chinook test DB used. I used a tutorial thru langchain to make this. I did have to go off tutorial tho to make it work where I initially made it. At the work place i made this we had a lot of firewall restrictions so i was not able to reach out to the langchain community hub to pull the script that was used. I ended up finding a script that was close and then made edits to it so i would get more accurate complete answers. I also have the LLM running offline thru Ollama!